+++ src/com/intellij/idea/plugin/hybris/diagram/typeSystem/node/TSDiagramNode.kt
@@ -36,11 +36,11 @@ class TSDiagramNode(val graphNode: TSGraphNode, provider: DiagramProvider<TSGrap
         .asSafely<TSGraphNodeClassifier>()
         ?.let {
             when (it.meta) {
-                is TSGlobalMetaEnum -> HybrisIcons.TypeSystem.ENUM
+                is TSGlobalMetaEnum -> HybrisIcons.TypeSystem.Types.ENUM
                 is TSGlobalMetaItem -> HybrisIcons.TypeSystem.ITEM
-                is TSGlobalMetaCollection -> HybrisIcons.TypeSystem.COLLECTION
-                is TSGlobalMetaMap -> HybrisIcons.TypeSystem.MAP
-                is TSGlobalMetaRelation -> HybrisIcons.TypeSystem.RELATION
+                is TSGlobalMetaCollection -> HybrisIcons.TypeSystem.Types.COLLECTION
+                is TSGlobalMetaMap -> HybrisIcons.TypeSystem.Types.MAP
+                is TSGlobalMetaRelation -> HybrisIcons.TypeSystem.Types.RELATION
                 else -> HybrisIcons.TypeSystem.FILE
             }
         }