+++ src/com/intellij/idea/plugin/hybris/impex/psi/references/ImpexFunctionTSItemReference.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -42,7 +42,7 @@ class ImpexFunctionTSItemReference(owner: ImpexParameter) : TSReferenceBase<Impe
         ?: super.calculateDefaultRangeInElement()
 
     override fun getVariants(): Array<LookupElement> = TSCompletionService.getInstance(element.project)
-        .getImpexInlineTypeCompletions(element.project, element)
+        .getImpExInlineTypeCompletions(element.project, element)
         .toTypedArray()
 
     override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {