+++ src/com/intellij/idea/plugin/hybris/psi/util/PsiUtils.java
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -43,6 +43,7 @@ public final class PsiUtils {
     private PsiUtils() {
     }
 
+    @Nullable
     public static Module getModule(final @NotNull PsiFile file) {
         final VirtualFile vFile = file.getVirtualFile();