+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/dto/CCv2BuildDto.kt
@@ -43,6 +43,7 @@ data class CCv2BuildDto(
         get() = CCv2Util.formatTime(startTime)
     val endTimeFormatted
         get() = CCv2Util.formatTime(endTime)
+    val duration = CCv2Util.getTimeDiffInMinutes(startTime, endTime).takeIf { it.toInt() != -1 } ?: "N/A"
 
     fun canDelete() = status != CCv2BuildStatus.DELETED && status != CCv2BuildStatus.UNKNOWN
     fun canDownloadLogs() = status != CCv2BuildStatus.DELETED && status != CCv2BuildStatus.UNKNOWN