+++ src/com/intellij/idea/plugin/hybris/toolwindow/system/type/tree/nodes/TSMetaEnumNode.kt
@@ -31,7 +31,7 @@ class TSMetaEnumNode(val parent: TSNode, meta: TSGlobalMetaEnum) : TSMetaNode<TS
 
     override fun update(project: Project, presentation: PresentationData) {
         presentation.addText(name, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        presentation.setIcon(HybrisIcons.TypeSystem.ENUM)
+        presentation.setIcon(HybrisIcons.TypeSystem.Types.ENUM)
         if (meta.isDynamic) {
             presentation.locationString = message("hybris.ts.type.dynamic")
         }